requirements/base.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary
python-decouple
